D3DFMIQ-737 Merged branches/FeatureBranches/NGHS_Framework14R to trunk
Former-commit-id: c5284945056ae9c4194674e5465a7a078b32a192
Former-commit-id: d2feede8205de05a3ae98add24875a5ca40e5ee2

+++ src/DeltaShell.Plugins.NetworkEditor/Import/OgrSchemaReader.cs
@@ -87,7 +87,7 @@ namespace DeltaShell.Plugins.NetworkEditor.Import
         private IList<string> GetFirstColumnAsValuesListFromLayer(Layer layer)
         {
             var lstValues = new List<string>();
-            int featureCount = layer.GetFeatureCount(1);
+            long featureCount = layer.GetFeatureCount(1);
             for (int i = 0; i < featureCount; i++)
             {
                 using (var row = layer.GetNextFeature())
